[development/directx_init_drawing_review] Move renderer backend to its own file Move renderer output window user data to its own file Move renderer info to its own file
Adjust includes accordingly
