+++ source/1_engine/rex_renderer_core/include/rex_renderer_core/commands/render_cmd.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "rex_engine/types.h"
-#include "rex_renderer_core/renderer.h"
+#include "rex_renderer_core/renderer_backend.h"
+#include "rex_renderer_core/resource_slot.h"
 
 namespace rex
 {
