+++ source/1_engine/rex_renderer_core/include/rex_renderer_core/renderer_info.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "rex_engine/types.h"
+
+namespace rex
+{
+    namespace renderer
+    {
+        struct Info
+        {
+            rsl::small_stack_string api_version;
+            rsl::small_stack_string shader_version;
+            rsl::small_stack_string adaptor;
+            rsl::small_stack_string vendor;
+        };
+    }
+}
