+++ source/1_engine/rex_renderer_core/include/rex_renderer_core/renderer_output_window_user_data.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "rex_engine/types.h"
+
+namespace rex
+{
+    namespace renderer
+    {
+        struct OutputWindowUserData
+        {
+            void* primary_display_handle;
+
+            s32 window_width;
+            s32 window_height;
+
+            s32 refresh_rate;
+
+            bool windowed;
+        };
+    }
+}
