+++ source/1_engine/rex_renderer_core/src/renderer.cpp
@@ -1,8 +1,16 @@
 #include "rex_renderer_core/renderer.h"
+#include "rex_renderer_core/renderer_backend.h"
 #include "rex_renderer_core/commands/render_cmd.h"
 #include "rex_renderer_core/resource_slots.h"
 #include "rex_renderer_core/log.h"
 #include "rex_renderer_core/shader_type.h"
+#include "rex_renderer_core/shader_platform.h"
+#include "rex_renderer_core/primitive_topology.h"
+#include "rex_renderer_core/index_buffer_format.h"
+#include "rex_renderer_core/viewport.h"
+#include "rex_renderer_core/scissor_rect.h"
+#include "rex_renderer_core/default_depth_info.h"
+#include "rex_renderer_core/default_targets_info.h"
 #include "rex_renderer_core/commands/begin_draw_cmd.h"
 #include "rex_renderer_core/commands/clear_cmd.h"
 #include "rex_renderer_core/commands/compile_shader_cmd.h"
